x/auction/client/cli/query_auction_pool.go
package cli

import (
    "context"
    "strconv"

	"github.com/cosmos/cosmos-sdk/client"
	"github.com/cosmos/cosmos-sdk/client/flags"
	"github.com/spf13/cobra"
    "github.com/Stride-Labs/stride/v5/x/auction/types"
)

func CmdListAuctionPool() *cobra.Command {
	cmd := &cobra.Command{
		Use:   "list-auction-pool",
		Short: "list all auction-pool",
		RunE: func(cmd *cobra.Command, args []string) error {
            clientCtx := client.GetClientContextFromCmd(cmd)

            pageReq, err := client.ReadPageRequest(cmd.Flags())
            if err != nil {
                return err
            }

            queryClient := types.NewQueryClient(clientCtx)

            params := &types.QueryAllAuctionPoolRequest{
                Pagination: pageReq,
            }

            res, err := queryClient.AuctionPoolAll(context.Background(), params)
            if err != nil {
                return err
            }

            return clientCtx.PrintProto(res)
		},
	}

	flags.AddPaginationFlagsToCmd(cmd, cmd.Use)
	flags.AddQueryFlagsToCmd(cmd)

    return cmd
}

func CmdShowAuctionPool() *cobra.Command {
	cmd := &cobra.Command{
		Use:   "show-auction-pool [id]",
		Short: "shows a auction-pool",
		Args:  cobra.ExactArgs(1),
		RunE: func(cmd *cobra.Command, args []string) error {
            clientCtx := client.GetClientContextFromCmd(cmd)

            queryClient := types.NewQueryClient(clientCtx)

            id, err := strconv.ParseUint(args[0], 10, 64)
            if err != nil {
                return err
            }

            params := &types.QueryGetAuctionPoolRequest{
                Id: id,
            }

            res, err := queryClient.AuctionPool(context.Background(), params)
            if err != nil {
                return err
            }

            return clientCtx.PrintProto(res)
		},
	}

	flags.AddQueryFlagsToCmd(cmd)

    return cmd
}
